+++ source/Application/Controller/Admin/NewsList.php
@@ -11,8 +11,7 @@ namespace OxidEsales\EshopCommunity\Application\Controller\Admin;
  * Admin news list manager.
  * Performs collection and managing (such as filtering or deleting) function.
  * Admin Menu: Customer Info -> News.
- *
- * @deprecated since v.5.3.0 (2016-06-17); The Admin Menu: Customer Info -> News feature will be moved to a module in v6.0.0
+ * @deprecated 6.5.6 "News" feature will be removed completely
  */
 class NewsList extends \OxidEsales\Eshop\Application\Controller\Admin\AdminListController
 {