+++ source/Application/Controller/Admin/NewsMain.php
@@ -15,8 +15,7 @@ use oxField;
  * Admin article main news manager.
  * Performs collection and updatind (on user submit) main item information.
  * Admin Menu: Customer Info -> News -> Main.
- *
- * @deprecated since v.5.3.0 (2016-06-17); The Admin Menu: Customer Info -> News feature will be moved to a module in v6.0.0
+ * @deprecated 6.5.6 "News" feature will be removed completely
  */
 class NewsMain extends \OxidEsales\Eshop\Application\Controller\Admin\AdminDetailsController
 {