+++ source/Application/Controller/Admin/NewsText.php
@@ -12,8 +12,7 @@ use stdClass;
 
 /**
  * Admin Menu: Customer Info -> News -> Text.
- *
- * @deprecated since v.5.3.0 (2016-06-17); The Admin Menu: Customer Info -> News feature will be moved to a module in v6.0.0
+ * @deprecated 6.5.6 "News" feature will be removed completely
  */
 class NewsText extends \OxidEsales\Eshop\Application\Controller\Admin\AdminDetailsController
 {